+++ api/tests/api/test_google_mail_notifications.rs
@@ -48,6 +48,7 @@ mod patch_resource {
             &settings,
             nango_google_mail_connection,
             None,
+            None,
         )
         .await;
 
@@ -106,6 +107,7 @@ mod patch_resource {
             IntegrationConnectionConfig::GoogleMail(google_mail_config.clone()),
             &settings,
             nango_google_mail_connection,
+            None,
             None,
         )
         .await;
@@ -175,6 +177,7 @@ mod patch_resource {
             &settings,
             nango_google_mail_connection,
             None,
+            None,
         )
         .await;
 
@@ -235,6 +238,7 @@ mod patch_resource {
             IntegrationConnectionConfig::GoogleMail(google_mail_config.clone()),
             &settings,
             nango_google_mail_connection,
+            None,
             None,
         )
         .await;