+++ api/tests/api/test_sync_google_mail_threads.rs
@@ -106,6 +106,7 @@ async fn test_sync_notifications_should_add_new_notification_and_update_existing
         &settings,
         nango_todoist_connection,
         None,
+        None,
     )
     .await;
     mock_todoist_sync_resources_service(
@@ -144,6 +145,7 @@ async fn test_sync_notifications_should_add_new_notification_and_update_existing
         IntegrationConnectionConfig::GoogleMail(google_mail_config.clone()),
         &settings,
         nango_google_mail_connection,
+        None,
         None,
     )
     .await;
@@ -360,6 +362,7 @@ async fn test_sync_notifications_of_unsubscribed_notification_with_new_messages(
         IntegrationConnectionConfig::GoogleMail(google_mail_config.clone()),
         &settings,
         nango_google_mail_connection,
+        None,
         None,
     )
     .await;