+++ api/tests/api/test_sync_todoist_tasks.rs
@@ -67,6 +67,7 @@ async fn test_sync_tasks_should_add_new_task_and_update_existing_one(
         &settings,
         nango_todoist_connection,
         None,
+        None,
     )
     .await;
     let todoist_projects_mock = mock_todoist_sync_resources_service(
@@ -280,6 +281,7 @@ async fn test_sync_tasks_should_add_new_task_and_delete_notification_when_disabl
         &settings,
         nango_todoist_connection,
         None,
+        None,
     )
     .await;
 
@@ -341,6 +343,7 @@ async fn test_sync_tasks_should_add_new_empty_task(
         IntegrationConnectionConfig::Todoist(TodoistConfig::enabled()),
         &settings,
         nango_todoist_connection,
+        None,
         None,
     )
     .await;
@@ -437,6 +440,7 @@ async fn test_sync_tasks_should_reuse_existing_sync_token(
         &settings,
         nango_todoist_connection,
         None,
+        None,
     )
     .await;
 
@@ -510,6 +514,7 @@ async fn test_sync_tasks_should_mark_as_completed_tasks_not_active_anymore(
         IntegrationConnectionConfig::Todoist(TodoistConfig::enabled()),
         &settings,
         nango_todoist_connection,
+        None,
         None,
     )
     .await;
@@ -629,6 +634,7 @@ async fn test_sync_tasks_should_not_update_tasks_and_notifications_with_empty_in
         IntegrationConnectionConfig::Todoist(TodoistConfig::enabled()),
         &settings,
         nango_todoist_connection,
+        None,
         None,
     )
     .await;
@@ -781,6 +787,7 @@ async fn test_sync_tasks_with_synchronization_disabled(
         &settings,
         nango_todoist_connection,
         None,
+        None,
     )
     .await;
     let todoist_mock = app.app.todoist_mock_server.mock(|when, then| {
@@ -820,6 +827,7 @@ async fn test_sync_all_tasks_asynchronously(
         IntegrationConnectionConfig::Todoist(TodoistConfig::enabled()),
         &settings,
         nango_todoist_connection,
+        None,
         None,
     )
     .await;
@@ -959,6 +967,7 @@ async fn test_sync_all_tasks_asynchronously_in_error(
         IntegrationConnectionConfig::Todoist(TodoistConfig::enabled()),
         &settings,
         nango_todoist_connection,
+        None,
         None,
     )
     .await;